# Performance optimization for size info in NTFS