I've found a major performance penalty in the space usage calculation of NTFS. Currenty the ClusterBitmap is read byte by byte from the underlying disk. This result in runtimes of up to one minute for large filesystems (>1TiB fs with a bitmap >50MiB).

I've changed the implementation to read the bitmap in blocks of 4K, which increased the performance by > 99% (in my test from ~1200ms down to ~45ms).

As a second optimization I've added a BitCounter which uses a precalculated lookup table to further speed up the size calculation from ~45ms down to ~15ms.